requirements.txt
astroid==2.3.3
autopep8==1.5.1
backcall==0.2.0
certifi==2020.6.20
colorama==0.4.3
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
future==0.18.2
inplace-abn==1.1.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.3.0
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.19.2
olefile==0.46
pandas==1.1.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
pip==20.2.4
prompt-toolkit==3.0.6
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pyzmq==19.0.2
scipy==1.5.4
seaborn==0.11.0
setuptools==50.3.0.post20201103
six==1.14.0
timm==0.3.1
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tornado==6.0.4
tqdm==4.51.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
wcwidth==0.2.5
wheel==0.35.1
wincertstore==0.2
wrapt==1.11.2
